Catalogue Shell scripts

+++ catalogue.sh
@@ -8,15 +8,15 @@ G="\e[32m"
 Y="\e[33m"
 N="\e[0m"
 
-LOGFILE=/tmp/$0-$DATE.log
+LOGFILE=/tmp/$(basename $0)-$DATE.log
 MONGODB_HOST=mongodb.mechanoidstore.online
 
 VALIDATE() {
     if [ $1 -ne 0 ]
     then
-        echo -e "$N $2 . . $R failed"
+        echo -e "$2 . . $R failed $N"
     else
-        echo -e "$N $2 . . $G sucess"
+        echo -e "$2 . . $G success $N"
     fi
 }
 
@@ -25,7 +25,7 @@ then
     echo -e "$R Root permissions required,$N run script with root user."
     exit 1
 else
-    echo -e "$G Script executed with root user."
+    echo -e "$G Script executed with root user. $N"
 fi
 
 dnf module disable nodejs -y &>> $LOGFILE
@@ -35,42 +35,59 @@ dnf module enable nodejs:18 -y &>> $LOGFILE
 VALIDATE $? "Enabled nodejs module version 18"
 
 dnf install nodejs -y &>> $LOGFILE
-VALIDATE $? "INSTALLATION nodejs"
+VALIDATE $? "Installed nodejs"
 
-useradd roboshop &>> $LOGFILE
-VALIDATE $? "roboshop user created"
+if id "roboshop" &>/dev/null; then
+    echo -e "$Y roboshop user already exists. $N"
+else
+    useradd roboshop &>> $LOGFILE
+    VALIDATE $? "Created roboshop user"
+fi
 
 mkdir -p /app
-VALIDATE $? "app directory created"
+VALIDATE $? "Created app directory"
+
+chown roboshop:roboshop /app
+VALIDATE $? "Set permissions for app directory"
 
 cd /app/
 
 curl -o /tmp/catalogue.zip https://roboshop-builds.s3.amazonaws.com/catalogue.zip &>> $LOGFILE
-VALIDATE $? "Download catalogue.zip"
+VALIDATE $? "Downloaded catalogue.zip"
 
-unzip /tmp/catalogue.zip &>> $LOGFILE
-VALIDATE $? "Unzipping catalogue.zip to app directory"
+unzip -o /tmp/catalogue.zip &>> $LOGFILE
+VALIDATE $? "Unzipped catalogue.zip to app directory"
 
 npm install &>> $LOGFILE
-VALIDATE $? "Downloading the dependencies"
+VALIDATE $? "Downloaded dependencies"
 
-cp /home/centos/roboshop-shell/catalogue.service /etc/systemd/system/catalogue.service &>> $LOGFILE
-VALIDATE $? "Copied the service file"
+if [ -f /home/centos/roboshop-shell/catalogue.service ]; then
+    cp /home/centos/roboshop-shell/catalogue.service /etc/systemd/system/catalogue.service &>> $LOGFILE
+    VALIDATE $? "Copied the service file"
+else
+    echo -e "$R Service file not found. $N"
+    exit 1
+fi
 
 systemctl daemon-reload &>> $LOGFILE
-VALIDATE $? "daemon-reload"
+VALIDATE $? "Reloaded systemd daemon"
 
 systemctl enable catalogue &>> $LOGFILE
-VALIDATE $? "Enable catalogue service"
+VALIDATE $? "Enabled catalogue service"
 
 systemctl start catalogue &>> $LOGFILE
-VALIDATE $? "Start catalogue service"
+VALIDATE $? "Started catalogue service"
 
-cp /home/centos/roboshop-shell/mongo.repo /etc/yum.repos.d/mongo.repo &>> $LOGFILE
-VALIDATE $? "Copied Mongo repo file"
+if [ -f /home/centos/roboshop-shell/mongo.repo ]; then
+    cp /home/centos/roboshop-shell/mongo.repo /etc/yum.repos.d/mongo.repo &>> $LOGFILE
+    VALIDATE $? "Copied Mongo repo file"
+else
+    echo -e "$R Mongo repo file not found. $N"
+    exit 1
+fi
 
 dnf install mongodb-org-shell -y &>> $LOGFILE
-VALIDATE $? "INSTALLATION MongoDB client"
+VALIDATE $? "Installed MongoDB client"
 
 mongo --host $MONGODB_HOST </app/schema/catalogue.js &>> $LOGFILE
-VALIDATE $? "Loading the data to MongoDB client"
+VALIDATE $? "Loaded data to MongoDB client"
